public/assets/python_function/requirements.txt
pycryptodome<=3.23.0
mysql-connector-python<=9.3.0
paramiko<=3.5.1